extract spec_cache
